+++ src/validation/functions/airtight/val_airtight_water_to_air.c
@@ -126,7 +126,7 @@
    }
    fpos = ftell(fi);
 
-   for (size_t i = 0; i < str_ncl; i++)
+   for (int i = 0; i < str_ncl; i++)
    {
      cylns[i] = str_cyln_alloc();
      str_cyln_read(fi, &fpos, origin, axis, &alpha, &radius, &height, base_s, 
@@ -154,11 +154,11 @@
    struct source ** srcs_2;
    srcs_2 = (struct source **) calloc(str_ncl, sizeof(struct source *));
 
-   for (size_t i = 0; i < str_ncl; i++)
+   for (int i = 0; i < str_ncl; i++)
    {
      srcs[i] = src_alloc();
      srcs_2[i] = src_alloc();
-     for (size_t j = 0; j < 3; j++)
+     for (int j = 0; j < 3; j++)
      {
        src_ref_o[j] = cylns[i]->o[j];
        src_s[j] = cylns[i]->s[j];
@@ -203,7 +203,7 @@
    printf("\n Emission at 0.0º, 0.0º --------------------------------------\n");
 
    mat_transpose(M_open, (double const **) cylns[ci]->top->M, 3, 3);
-   for (size_t i = 0; i < 3; i++)
+   for (int i = 0; i < 3; i++)
    {
      ref_o_2[i] = -cylns[ci]->h * cylns[ci]->top->M[i][2];			
      ref_o_1[i] = -cylns[ci]->o[2] * cylns[ci]->M[i][2];			
@@ -215,7 +215,7 @@
      if ( srcs[ci]->rotate_f )
      {
        ROT_VEC_UNIT(u_rot, ray.u, srcs[ci]->M);
-       for (size_t i = 0; i < 3; i++)
+       for (int i = 0; i < 3; i++)
          ray.u[i] = u_rot[i];
        COS_TO_SPH_UNIT(ray.s, ray.u, random);
      }
@@ -233,7 +233,7 @@
    printf("\n Emission at random ------------------------------------------\n");
 
    mat_transpose(M_open, (double const **) cylns[ci]->top->M, 3, 3);
-   for (size_t i = 0; i < 3; i++)
+   for (int i = 0; i < 3; i++)
    {
      ref_o_2[i] = -cylns[ci]->h * cylns[ci]->top->M[i][2];			
      ref_o_1[i] = -cylns[ci]->o[2] * cylns[ci]->M[i][2];			
@@ -246,7 +246,7 @@
      if ( srcs[ci]->rotate_f )
      {
        ROT_VEC_UNIT(u_rot, ray.u, srcs[ci]->M);
-       for (size_t i = 0; i < 3; i++)
+       for (int i = 0; i < 3; i++)
          ray.u[i] = u_rot[i];
        COS_TO_SPH_UNIT(ray.s, ray.u, random);
      }
@@ -282,7 +282,7 @@
    printf("\n Emission at 0.0º, 0.0º --------------------------------------\n");
 
    mat_transpose(M_open, (double const **) cylns[ci]->top->M, 3, 3);
-   for (size_t i = 0; i < 3; i++)
+   for (int i = 0; i < 3; i++)
    {
      ref_o_2[i] = -cylns[ci]->h * cylns[ci]->top->M[i][2];			
      ref_o_1[i] = -cylns[ci]->o[2] * cylns[ci]->M[i][2];			
@@ -294,7 +294,7 @@
      if ( srcs[ci]->rotate_f )
      {
        ROT_VEC_UNIT(u_rot, ray.u, srcs[ci]->M);
-       for (size_t i = 0; i < 3; i++)
+       for (int i = 0; i < 3; i++)
          ray.u[i] = u_rot[i];
        COS_TO_SPH_UNIT(ray.s, ray.u, random);
      }
@@ -312,7 +312,7 @@
    printf("\n Emission at random ------------------------------------------\n");
 
    mat_transpose(M_open, (double const **) cylns[ci]->top->M, 3, 3);
-   for (size_t i = 0; i < 3; i++)
+   for (int i = 0; i < 3; i++)
    {
      ref_o_2[i] = -cylns[ci]->h * cylns[ci]->top->M[i][2];			
      ref_o_1[i] = -cylns[ci]->o[2] * cylns[ci]->M[i][2];			
@@ -326,7 +326,7 @@
      if ( srcs[ci]->rotate_f )
      {
        ROT_VEC_UNIT(u_rot, ray.u, srcs[ci]->M);
-       for (size_t i = 0; i < 3; i++)
+       for (int i = 0; i < 3; i++)
          ray.u[i] = u_rot[i];
        COS_TO_SPH_UNIT(ray.s, ray.u, random);
      }
@@ -361,7 +361,7 @@
    printf("\n Emission at 0.0º, 0.0º --------------------------------------\n");
 
    mat_transpose(M_open, (double const **) cylns[ci]->top->M, 3, 3);
-   for (size_t i = 0; i < 3; i++)
+   for (int i = 0; i < 3; i++)
    {
      ref_o_2[i] = -cylns[ci]->h * cylns[ci]->top->M[i][2];			
      ref_o_1[i] = -cylns[ci]->o[2] * cylns[ci]->M[i][2];			
@@ -373,7 +373,7 @@
      if ( srcs[ci]->rotate_f )
      {
        ROT_VEC_UNIT(u_rot, ray.u, srcs[ci]->M);
-       for (size_t i = 0; i < 3; i++)
+       for (int i = 0; i < 3; i++)
          ray.u[i] = u_rot[i];
        COS_TO_SPH_UNIT(ray.s, ray.u, random);
      }
@@ -391,7 +391,7 @@
    printf("\n Emission at random ------------------------------------------\n");
 
    mat_transpose(M_open, (double const **) cylns[ci]->top->M, 3, 3);
-   for (size_t i = 0; i < 3; i++)
+   for (int i = 0; i < 3; i++)
    {
      ref_o_2[i] = -cylns[ci]->h * cylns[ci]->top->M[i][2];			
      ref_o_1[i] = -cylns[ci]->o[2] * cylns[ci]->M[i][2];			
@@ -404,7 +404,7 @@
      if ( srcs[ci]->rotate_f )
      {
        ROT_VEC_UNIT(u_rot, ray.u, srcs[ci]->M);
-       for (size_t i = 0; i < 3; i++)
+       for (int i = 0; i < 3; i++)
          ray.u[i] = u_rot[i];
        COS_TO_SPH_UNIT(ray.s, ray.u, random);
      }
@@ -439,7 +439,7 @@
    printf("\n Emission at 0.0º, 0.0º --------------------------------------\n");
 
    mat_transpose(M_open, (double const **) cylns[ci]->top->M, 3, 3);
-   for (size_t i = 0; i < 3; i++)
+   for (int i = 0; i < 3; i++)
    {
      ref_o_2[i] = -cylns[ci]->h * cylns[ci]->top->M[i][2];			
      ref_o_1[i] = -cylns[ci]->o[2] * cylns[ci]->M[i][2];			
@@ -451,7 +451,7 @@
      if ( srcs[ci]->rotate_f )
      {
        ROT_VEC_UNIT(u_rot, ray.u, srcs[ci]->M);
-       for (size_t i = 0; i < 3; i++)
+       for (int i = 0; i < 3; i++)
          ray.u[i] = u_rot[i];
        COS_TO_SPH_UNIT(ray.s, ray.u, random);
      }
@@ -469,7 +469,7 @@
    printf("\n Emission at random ------------------------------------------\n");
 
    mat_transpose(M_open, (double const **) cylns[ci]->top->M, 3, 3);
-   for (size_t i = 0; i < 3; i++)
+   for (int i = 0; i < 3; i++)
    {
      ref_o_2[i] = -cylns[ci]->h * cylns[ci]->top->M[i][2];			
      ref_o_1[i] = -cylns[ci]->o[2] * cylns[ci]->M[i][2];			
@@ -482,7 +482,7 @@
      if ( srcs[ci]->rotate_f )
      {
        ROT_VEC_UNIT(u_rot, ray.u, srcs[ci]->M);
-       for (size_t i = 0; i < 3; i++)
+       for (int i = 0; i < 3; i++)
          ray.u[i] = u_rot[i];
        COS_TO_SPH_UNIT(ray.s, ray.u, random);
      }
